removed e.stopPropagation call in keyStop function
Hi, first of all I want thank you for good library :).
I am using it in my application and sometime I need to check
is Esc-button presed and do some processing. But becaouse of
stopPropagation function I could not to know about it.
Is it realy necessary? Can we just drop it?
Or exclude in some situations?